Simulate IAAS-175
It is possible for bbslave to fail to converge if the specified
configuration is impossible. This can happen, for example, if
the desired app version does not match the actual version of the
package specified by the URL.

This change simulates this by making the mock pf9app extract the
actual version from the package name, if it is specified.
Modified the unit test to simulate this scenario.

This is in preparation to actually solving IAAS-175.