Makefile for resource manager
This makefile currently supports building the resource manager pre-reqs (for
testing purposes) and building the resmgr rpm. Introduced a dummy init.d script
as well for now, but that will change based on how we decide to bring up the
pecan app.

Testing done: make pre-reqs, make clean and make rpm. Used rpm -qlp to verify
the contents of the rpm.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/215/)