Introduce bbmaster and resmgr customization scripts in pf9-du
Package the corresponding customization scripts in the bbmaster and resmgr rpms.
1. customize-bbmaster sets the AMQP broker host address.
2. customize-resmgr sets the SQL connect string. Since RDS setup is not yet in
place, this script reads the nova db connection string and edits it to use the
test db instance in the nova DB.
3. Start both services at the end of each customization script.

In the new scheme of customization, the input arguments are passed through
environment variables.

Testing done: Built and installed the rpms. Manually ran the new customization
scripts and verified they work fine. Need to wait for the rest of the framework
to be ready before I can do a complete integration test.

Reviewed by: leb, hli (https://rbcommons.com/s/platform9/r/228/)