Check keystone permissions in the resource manager
This change integrates the keystone authentication middleware into the resource
manager. If the server is started with the new paste configuration and server startup
script, the server will require a valid authentication token in the X-Auth-Token
header, and for PUT/DELETE, it requires that the token user has the 'hostadmin' role
on the tenant.

The server can still be run with the old pecan config, and in this case, the middleware
doesn't appear in the wsgi pipeline and no keystone interaction is required.

Unit tests exercise the paste-configured wsgi app, with a mocked (httpretty) keystone,
and verify that 401s and 403s are returned as appropriate. Old api tests run without auth
as before.

testing: tested manually with curl and a real keystone, unit tests with a fake keystone.