Make nova AMQP communication SSL based - part 2
This is the pf9-du side changes for the SSL based AMQP communication on nova
side.

We now create the pf9 user and pf9group group as part of the hostagent rpm. The
root user is added to the pf9group. It needs access to the certs and so does the
pf9 user. The certs are now owned by the pf9group group. The hostagent rpm also
provides a pf9-bbslave capability which is currently leveraged by ostackhost
rpm.

Removed the rabbitmq password from the backbone config files. It is part of the
code now.
Also, fixing the customize-hostagent-rpm script which was moved from pf9-main
and
needs to be committed as part of this change.

Bonus: bbslave logging was broken because of some code optimization that was
done incorrectly as part of my previous change. Fixed it now.

Testing done: Built the hostagent rpm. Installed it on a KVM host. Verified
perms of the cert files. Also verified ostackhost is able to talk to the DU.

Reviewed by: leb(https://rbcommons.com/s/platform9/r/225/)