Create hostagent-tarball rpm
The rpm will contain the hostagent tarball and the hostagent customization
script. The customization script is heavily borrowed from the
customize-agent-rpms script in pf9-main/scripts/du-customize, along with
ostackhost rpm references being removed.

The assumption is that this customization script will run after the certificates
have been generated and before bbmaster and resmgr comes up.

Testing done: Ran make ha-tarball-rpm. Copied the rpm over to DU machine,
installed rpm, verified that the tar.gz file is placed in the correct dir. Ran
the script manually and verified that the hostagent rpm was staged in the
expected static directory.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/230/)