Setup log rotation for host agent
Use the python rotating file logger to setup the logging for the host agent.
Opted for this instead of logrotate service since it may not be a good idea to
assume that the host machines have logrotate setup.

Optionally, do console based logging for tests only so that the tests dont end
up dropping log files on various test machines

This addresses IAAS-176

Testing done : Installed the hostagent. Edited the config file to have small file size
and pumped log messages into the logger. Ran unit and integration test for
backbone.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/240/)