Fixup pf9-du components to work with new DU deployment scheme
This change covers multiple modules that are part of the DU.
1. Made pf9cert to be part of bbmaster rpm. During certificate generation, we
will use the /opt/pf9/bbmaster/bin/python executable to generate the
certificates. Also, redid gencerts.py to use optparse and moved it under
bbmaster dir for now.
2. nginx server is being used to serve out the hostagent rpm. Fix the hostagent
rpm location based on what it expects. Also, remove the workaround done for the
horizon.conf for apache server.
3. Setup the pf9-bbmaster init.d script. This is more of a hack. Created
/opt/pf9/bbmaster/bin/pf9-bbmaster as a symlink to /opt/pf9/bbmaster/bin/python.
Then this is used to make the init.d script handle start, stop and status
correctly. Internally, it is running pecan serve config.py as a background
process. This will be fixed soon to act like a true daemon. Note that I cannot
directly kick off the python executable because the init.d stop may kill all
python processes based on how killproc works.
4. Fix a bug in the customize-resmgr script. sys.exit causes an exception to be
raised, which was being caught as an error even if it was sys.exit(0)

Note that there are a couple of hacks (which I am definitely not proud of) in
here.
1. pf9cert needs a proper home. Right now it is piggyback-ing on the bbmaster
rpm.
2. pf9-bbmaster init.d script needs to have a better implementation.

Testing done: Built all the rpms. Deployed a new AMI, pushed these rpms on that AMI.
Reinstalled these rpms and ran the customize script. Verified the result.

Reviewed by: rparikh (https://rbcommons.com/s/platform9/r/247/)