Make generate certs handle spaces in customer full name
Got rid of the shell script that was driving the cert generation. Instead of
that, we now have the python script driving all the cert generation. The script
also looks at the environment variables for values for customer short and full
name. I have added a check after options parsing to ensure these "mandatory"
attributes are not none.

Rest of the code is as is from before. However, reviewboard seems to
misunderstand how the diff should be handled. It is treating the diff as a
rewrite of the shell code instead of a file rename.

This addresses IAAS-244 and part of IAAS-233.

Testing done: Ran the script with the customer full name having spaces and
ensured the certs are generated.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/265/)