Adding python-setuptools dependency to host agent rpm spec
The use of --system-site-packages while building the hostagent rpm is causing
some different behavior of how the hostagent rpm packages setuptools in its
venv. Made the rpm now depend on python-setuptools.

This should address IAAS-251.

Testing done: This bug was hard to repro because it seems to happen mainly with
the rpms coming out of teamcity. I tried to build the rpm into my DU and setup
on the machine where this was failing. It succeeds. However, this is not a pure
test. Will test again once it is in teamcity.

Reviewed by: sraghuram, leb (https://rbcommons.com/s/platform9/r/275/)