Host removal in resource manager & backbone polling for host info
Hosts are broadly classified as authorized and unauthorized (no roles assigned).
In case of unauthorized hosts, the resource manager stores its state only in
memory. It shall drop it from this state if the host has not sent a status
message within the expected time interval (300s specified in config file). If
authorized, then the host shall be marked as inactive in the DB, if it has not
received a status message within the expected time interval.
Exposed a new endpoint DELETE /v1/hosts/__id__. A delete operation on an
unauthorized host is a no-op. In case of an authorized host, we remove all the
assigned roles to the host and then remove the host from our persistence store.
This is the way we expect users to remove the host from their inventory. If the
hostagent on such a deleted host continues to send status messages, these hosts
are classified as unauthorized hosts now.

This change also implements a backbone master poller, which queries for the
state of the hosts. The poller first classifies hosts reported by backbone into
new hosts, existing hosts and deleted hosts. New hosts, if within the active
host time interval are added to the unauthorized list. Deleted hosts are marked
as inactive if they were previously authorized, else removed from the
unauthorized host list. Existing authorized hosts are first marked
active/inactive based on their status times. Convergence is attempted if we
detect that the hosts are not in the expected app configuration. In case of
unauthorized hosts, we update the status timestamp of the host.

Backbone master changes:
Backbone master now reports the timestamp when a host status message was
received. This is used by resource manager to make decisions on the
active/inactive state of the host.

Other fixes:
- Added logging for resource manager. This leverages pecan logging configuration
capabilities.
- Made notifier work with the default password if it is not specified in the
config file.
- Moved all the resources endpoints to hosts endpoints. For example
/v1/resources is now v1/hosts.

Note: Some settings in current test config files were tweaked to make it easier
for me to test my changes. I will revert these once I have some reviews and some
testing in place.

Testing done: Ran the existing unittests for bbmaster and resmgr. Lot of manual
testing. Added the hostagent to 2 hosts. Verified they are detected in the right
order. Performed a DELETE host on one of them. Verified it was a no-op. Pushed
ostackhost role on to both of them. Deleted the ostackhost role in one of them
using the DELETE role API. Then performed a DELETE host on the other. Verified
that the ostackhost was cleaned up in each case. Also, verified that the
notifier logged changes being reported to AMQP broker.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/298/)