Run 50-customize-resmgr only if no rds setup
I'm adding a new script to the pf9-main du-customize stuff that
creates a remote database for the resource manager on an rds instance
and then starts the service. For dev puposes, it's nice to be able
to build a DU with a local mysql, so rather than removing this
script altogether, I'm skipping it when the rds configuraiton
(DBENDPOINT) is passed in the environment.

testing: tested on a recent du AMI with and without rds