Add heartbeats to notifier library.
On the web client side, there are issues that combine to
break our notification system (and its authentication mechanism):

1. Chrome has a bug with websockets not closing even when the
   server side socket has closed. They just hang in a bad state
   forever.

2. The stomp-websocket library which runs over the websocket can
   solve (1) by using STOMP protocol heartbeats, and explicitly
   disconnecting the websocket after detecting a period of
   server inactivity.

3. Unfortunately, on the server side, rabbitmq's Web Stomp plugin
   does NOT support STOMP heartbeats (see the bottom of this
   article: http://www.rabbitmq.com/web-stomp.html)

So, the proposed solution is to add application-level heartbeat
messages to the python notifier library. This will enable
stomp-websocket to detect inactivity after the connection has been
closed on the server side.

I tested this manually with Chrome & Firefox on the client side, and
ngnix, pf9-authproxy, and resource manager (in mock mode) on the
server side.