Implementation of host agent update + misc fixes
Added 2 new REST API to deal with host agent update in backbone.

1. GET /v1/hosts/__id__/hostagent: will return the host agent details. Currently
I am considering reporting status and version of the host agent. Status could be
running or in case of updates, report a status of 'updating' prior to the actual
update.
2. PUT /v1/hosts/__id__/hostagent: This upgrades the host agent on a particular
host. This takes in a name, version, URL in the request body. The URL is from
where the host agent will be downloaded and updated on the host.

The bbmaster will pass a new message 'set_agent' through rabbitmq to the slave
where the host agent is being updated. The host agent shall deal with updating
itself. Currently, the update is going to be in place. Side by side update will
come in later.

On the host agent side, introduced some agent specific classes in pf9app layer.
These are similar to the other Pf9App class, except for some special
capabilities like updating a package or querying the host agent package. On
receiving an update message, the agent responds back with an 'updating' agent
message to the master. On success of the update, the host agent shall restart
and the new version of the host agent will send a status message with 'running'
status and new version of the host agent. If the update failed, then the host
agent sends back a message with 'running' status with the older version of the
agent.

Introduced a du_ctl script. This shall be used to interact with the bbmaster
REST endpoints and control pushing new agent updates to a host. This script
currently does only host agent updates. It can (1) list the hosts and their
agent inf0, (2) update the agent. Plan is to evolve this script to do other du
operations like mark roles active, etc. A TODO here is to find a home for this
script (probably the pf9-du rpm?)

Note, that there is a case where the du_ctl script may fail to detect a host
agent failure, because bbmaster doesn't maintain status history. The du_ctl will
just timeout in such cases. Hopefully, we will never have a hostagent update
fail :). Making this work the right way may involve a more intensive change in
backbone. Will use a JIRA item for that purpose.

Also adding in a bug fix, where yum API is not behaving as we expect for
install/uninstall. Using yum CLI instead.

Testing done: Ran the update through the ctl script. Ensured it succeeds. Ran
the unit-tests for bbslave, bbmaster and bbone. All of them pass. Ran host
authorize/remove from the UI as well.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/321/)

Issues addressed: IAAS-371, IAAS-198