Better DB session management in resource manager
Introducing DB session management using contextmanager. When the session is
going out of scope, it is inspected to check if there are pending uncommitted
changes (new, modified, deleted entries). If present, they are committed or
rolled back in case of error. Finally, the session is closed.

The change also takes care of all the methods interacting with the DB to use the
new way of obtaining the DB session. One unexpected change here is that
query_host* will return a dict object instead of the ORM object. This is because
Host.roles is an external reference (SQL join) for the Host ORM object. This is
a lazy reference in SQLAlchemy and cannot be accessed once outside the scope of
the session.

The backbone poller now catches all exceptions so that it doesn't go down for
temporary issues.

Also, fixed HostInventoryMgr.get_host() to return a dict of just host attrs
instead of { host_id : {host attrs} }, for consistency purposes. Fixed all the
relevant call sites.

Note that this is probably not the best way we want to maintain DB session
scope. This will be revisited after discussion with others.

Bonus fix: Add bbmaster to chkconfig of the DU as part of customization script.

Testing done: From the UI, ran through the add and authorize host sequence.
Verified that the backbone poller was running fine after a few hours, with the
connection pool reporting sane numbers. Built a script to send 1000 get
/v1/hosts and 1000 /v1/hosts/__id__ requests to resource manager. All requests
succeeded.