Remove host agent config cache during uninstall/upgrade
Remove host agent config cache during uninstall/upgrade. If we leave it around,
then the subsequent host agent install will look up this cache on service
startup and try to converge app state. There is a good chance this type of
convergence may fail.

Testing done: Built the rpm. Setup the rpm in the DU. Installed it on a host and
removed it. Ensured the directory is cleaned up.

Reviewed by: hli (https://rbcommons.com/s/platform9/r/353/)