Fixup the host remove algorithm
Previously, the resource manager would push the role remove configuration to the
host and then proceed to register these changes in the database. The issue is
that between the time this push happens and the DB gets updated, the backbone
poller thread may read the old configuration from the DB and try to push that
back to the host. This leads to 2 issues. (1) ostackhost gets reinstalled (2)
the feedback time for the UI/client to notice that the server has registered
this change can be long.

Fix for this is to leverage the current backbone design that the operations on
the slave are idempotent. So, we will fix register the changes in the DB, notify
the clients through notifier that a change has happened. And then push the
changes to the backbone to fix up the configuration. Since the changes are
registered in the DB first, further pushes from backbone poller will result in
the same configuration pushed down to backbone. This should not cause the
ostackhost to get reinstalled.

A future enhancement will be to deal with status reporting (success, failures)
around this operation.

Bonus fix: Fixed up some logging in hostagent. There were some logs that were
not getting published at all.

Testing done: Through the UI, performed about 20-25 authorize/delete cycles on 2
hosts. The host always ended up in the expected state. The changes on the UI on
delete operation was noticeably quicker.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/388/)

Issues addressed: IAAS-325