Add log rotation for backbone master
Add log rotation for backbone master. Leverage the pecan logging integration for
bbmaster. Also, moved the logs from /var/log to /var/log/pf9 to make it
consistent with resmgr.

Testing done: Installed the new rpm in the DU and verified the logs are generated in the
expected place and rotation does happen.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/397/)

Issues addressed: IAAS-295