Make hostagent rpm version & release setting bit more flexible
This change makes the host agent version and release string configurable through
the host agent customization script. If HOST_AGENT_VERSION variable is defined,
the customization script will use it instead of the baked in version. If
HOST_AGENT_RELEASE variable is defined, it will be used instead of the baked in
release. This version, release is then passed into the hostagent_rpmbuild.sh and
eventually into the hostagent rpm spec.

The intent here is to make it easier to generate a new hostagent rpm as part of
the integration tests. This is needed to test host agent update code.

Testing done: Deployed a new hostagent-tarball-rpm in the DU. Ran the host agent
customization script. Verified a new rpm was generated with the expected
symlink. Ran the customization script again with the HOST_AGENT_VERSION,
HOST_AGENT_RELEASE variables and verified the new version of the rpm was
generated and the symlink was pointing to the new rpm.

Reviewed by: leb, hli (https://rbcommons.com/s/platform9/r/460/)