Fix notification for unauthorized host deletion.
I noticed that the UI was not showing the deletion of an unauthorized
hosts whose pf9-hostagent had been deleted for a long time.
The bug was in resmgr.
I also decreased the unauthorized host inactivity timeout to 90 seconds,
which is 3 hostagent heartbeats. 5 minutes seems too long.

Testing: continually installed and uninstalled pf9-hostagent from a
host, and verified that the UI eventually reflected the presence/absence
of the unauthorized host.