Change resmgr configuration apply to be non-blocking
In the code where we push configuration changes to the backbone, we had 2 issues
with the implementation logic

1. The push call would check if the backbone state was the same as what was
being pushed. If the same, it would not push any configuration changes.
Considering backbone is idempotent and can deal with multiple messages with same
configuration, this check in resmgr did not add much value. It did lead to
issues when the resource manager did a role add and delete in quick succession.
The backbone had not yet converged on the add and the subsequent delete was
dropped because of this particular check.

2. There was a poll loop that would wait to ensure that the backbone changes had
in fact converged, prior to returning from the push call. This was probably not
the right way to do this. The REST API call to add a role to the host could
block for a very long time because of this, which is not how we want the REST
API to behave. This was leading to the web server returning 504. Reporting
success or failure on convergence of backbone is a separate work item wrt task,
event reporting infrastructure.

Testing done: Ran multiple runs of the integration test. They passed and no 504s
were seen. Through the UI, multiple times ran the authorize and remove host
operations in quick succession. Noticed that the operations queued up and
eventually converged with the host to make the host unauthorized.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/472/)

Issues addressed: IAAS-381, IAAS-414