Introduce du-tools rpm
Currently, this rpm only holds the du_ctl script. In future, additional
analytics, troubleshooting software can be added into this rpm. Since it is just
one python script which can be run standalone (its dependencies are satisfied by
the global python installation of the DU), I did not go to build virtual
environment. Also, used simple rpmbuild for now to keep this rpm build quick and
simple.

This rpm should eventually make its way to our yum repo through teamcity builds.
The intention of the rpm is to be installed in the DU, as it is needed. Once
this is in the yum repo, the integration tests will install the rpm and initiate
a host agent update.

Testing done: Built the rpm and verified its contents with rpm -qip --scripts
<rpmname>

Reviewed by: leb, rdeuel (https://rbcommons.com/s/platform9/r/473/)