Explicitly show in log if a customization sub-process takes too long.
We've see occasional failures in nightly "pf9.platform9.net" deployment jobs
in which no error message is given, and the job got killed after 2 hours.
Recent examples are build #63, #66, #68. I suspect that this was caused by
a hung customization process, 70-user-setup.py, specifically.

This change limits each customization sub-process to 5 minutes, by which
time a SIGTERM will be sent to it. If the process doesn't die in 5 minutes,
a SIGKILL is sent. The log file will clearly show all these.