Generate UUID for host_id instead of using MAC address
See IAAS-460 for info about the problems with using MAC.
The generated UUID is stored in the file system if not found,
and persists, even after pf9-hostagent removal. A user
can manually remove the file if he wants the host to get a new UUID.

The file is /etc/pf9/host_id.conf
For historical reasons, a copy of the effective host_id is also
saved under /var/opt/pf9/hostagent/data.conf for pf9-ostackhost.
That one is left alone to preserve compatibility with
pf9-ostackhost and also because that directory gets cleaned
up after pf9-hostagent removal.

Upgrade implications:
This change fundamentally changes the way hosts are identified,
so it may not work well with in-place hostagent upgrade. It
might work but the host would show up as a new host.

Other changes:

1. Fixed bugs in 2 Makefiles that masked unit test errors (!!!).

2. I noticed that the bbslave unit test, despite passing, showed
strange output in the logs, with some messages repeated several
times. I found two problems:

- Each test spawns a new thread, while the old one from the previous
  test is still around. For reasons I don't fully understand yet,
  it looks like the old thread is the one that always processes messages
  sent by the new test, while the new thread isn't getting any.
  I fixed this by terminating the thread after each test. Since Python
  doesn't provide a mechanism to kill a thread, I had to add an 'exit'
  message to ask the thread to exit cleanly.

- Each test's new thread runs reconnect_loop(), which keeps adding new
  appenders to the existing logger, resulting in duplicate log lines.
  I fixed by only adding new appenders if none exist.

Testing:
- bbslave unit-test
- bbone integration test
- On a host without the UUID file, verified it is generated the
  first time, and reused afterwards.
- Will run integration test