Support for multiple roles in resmgr
Instead of hard coded role information in resource manager, load the role
version from role metadata files that will be provided by the role (wrapper) rpm
itself. These metadata files will be placed in
/etc/pf9/resmgr_roles/<role_name>/<role_version>/<filename>.json. When resmgr
starts, it will read the role files present in the dir and setup the role
configuration data. This should make the loading of multiple roles dynamic. If
loading a role fails because of bad metadata format, that role metadata is
skipped and we continue to the next one role file.

For substituting attributes in the configuration with runtime values, the name
values can be fed from the environment variables. In addition to this, there
will be a predefined list of name, values that will be maintained. Note that
location of this predefined list is not clear yet. The location and what these
predefined values will contain will be decided soon.

Fixed up the pf9-testapp so that I could use it to validate the multiple role
capability on my test setup.

Testing done: Patched the resmgr in the DU. Ran customization on the DU.
Manually setup the testapp in the DU. Installed the testapp role along with
ostackhost role and verified that it works.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/519/)

Issues addressed: IAAS-407