Catching error when calling stop on pf9-resmgr service
When calling stop multiple times on this service, an error was being thrown and a stack trace was being printed. This commit catches the error and does not print any additional information.

Testing done: Tested the command "service resmgr stop" on a DU. There was no output after running it many times.
Also tested that a successful stop resulted in an exit code of 0 while an unsuccessful stop returns exit code 1.

Reviewed by: leb, pshanbhag (https://rbcommons.com/s/platform9/r/520)

Issues addressed: IAAS-441