Move /etc/init.d/pf9-hostagent to /etc/rc.d/init.d/pf9-hostagent
Since Red Hat uses a sym link from /etc/init.d to /etc/rc.d/init.d, this update
changes the destination directory for bbmaster and bbslave services from
/etc/init.d to /etc/rc.d/init.d.

Testing done: The new hostagent-tarball-rpm was generated and installed on a DU.
A kvm instance was created on vSphere, and the hostagent rpm was downloaded on
the kvm instance from the DU. The hostagent rpm was then installed. It was
verified that the DU could authorize the host and that the hostagent was running
on the kvm host.

Reviewed by: hli, pshanbhag (https://rbcommons.com/s/platform9/r/526/)

Issue resolved: IAAS-461