Role version update support in resource manager
This change deals with supporting updates of role versions in resource manager.
For a role update, the role rpm needs to include updated version of the role
metadata along with the updated bits. On reload of resource manager, all the
metadata files are read in and active version for each role is determined as
below:
  - if the role was not known to resource manager before (not in the DB), it
would mark the highest version of the role as the active version of the role
  - if the role was previously known to resource manager (was in the DB), that
version continues to remain the active version. The new version can be marked
active using the new REST API in resource manager.

Introduced a new REST API in resource manager to mark a particular version of
the host as active. It is a PUT call on the /v1/roles/__role_name__ endpoint,
with the version to be marked active passed in the request body.

The du_ctl script was also update to provide 2 more commands.
   - Change a role to mark a different role version as active. This just invokes
the above PUT call.
   - Apply a role to a host. This invokes the role assignment PUT call on
resource manager. A side effect of this is that a host can be updated to a new
(active) version of the role by issuing this command.

Here is the workflow we expect to follow when the active version of a role
changes
   - Hosts that did not have this role previously assigned will be assigned this
new active version of the role whenever the role is applied to these hosts.
   - Hosts that had a previous version of the role assigned, will continue to
remain at that previous version. No automatic update of the role for these
hosts. To move to a new version of a role, the role assignment PUT call should
be invoked again on these hosts. The call will check that the host doesn't have
the active role version and perform a role update operation on the host.

Note: the du_ctl script assumes we have access to a keystone token. This will
probably change once we figure out our DU service account story (ongoing mail
thread).

Testing done: Updated a DU with my changes. Assigned a host with the 1.0.0
version of the ostackhost. Locally built a new higher version of ostackhost rpm
wrapper. Updated the wrapper rpm in the DU.
On restart of resource manager, noticed that 1.0.0 remained the active version
in the DB. The host also remained at ostackhost 1.0.0. Used the du_ctl to switch
the active version of the role. Verified in the DB that the active version did
change. Applied the role to the host using the du_ctl command and verified that
the host got the updated role version.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/539/)