Restart services instead of start when migrating db to rds
Currently, the customization step begins as soon as sshd is up. It then stops
the nova service, migrates the db to rds, and starts the nova service. However,
nova could have already been started, in which case start would be a no-op. This
change ensures nova restarts after migrating the database. However, nova and
other services could still interfere with the migration of the database. See
IAAS-519 for proposed solutions.

Testing done: Successfully ran customization on a du instance with an rds
endpoint

Issues addressed: IAAS-483