Fix restart service for resmgr
The python-daemon library does not allow stop (or restart) to be called when the
process is already stopped because of some pid file issue which throws an error.
This update changes the restart service to call stop then start, instead of
calling restart on the library. From IAAS-441, calling stop on a stopped resmgr
will ignore the error.

Issues addressed: IAAS-520

Testing done: Launched Demo DU with an ami that contained the changes. Ensured
resmgr was running when the DU loaded. Also ensured calling restart on the
resmgr starts the service even though it was stopped before.