Add pf9-imagelibrary-wrapper dependency on pf9-du + resmgr bug fix
The pf9-imagelibrary role now needs to be included in the DU. So, add a
dependency in the uber rpm.

While testing this new role in resource manager, noticed that the active role
version determination logic was not identifying active role versions at times.
This was because of a break instead of continue statement, which was preventing
processing of all identified roles.

Testing done: Authorization of image role from resmgr.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/608/)