Fix upgraded cert permissions.
When the hostagent is upgraded, hostagent files are reverted to the default
permissions. This commit changes the group of the hostagent files to pf9group
after a hostagent update.

Issues addressed: IAAS-538

Testing done: Launched a DU and a host. Installed hostagent and ensured the UI
was waiting for authorization. Installed the new hostagent tarball on a du and
customized it to produce a new hostagent rpm. Downloaded the new hostagent on
the host and updated it. Authorized the host using this new hostagent and
ensured the UI reflected the changes.