Deal with simultaneous role assignment for a host
There was a race condition in the ORM layer of resource manager when multiple
roles get assigned simultaneously to an unauthorized host. This occurs when we
attempt upsert equivalent workflows in SQLAlchemy. An unauthorized host when
assigned a role, is added into the resgmr database. When we have 2 REST calls
made for assigning roles to the same host, we have 2 threads trying to add the
host into the database. In both threads, the ORM tries to read the DB and
determines the host entry should be an insert (not update). When the threads try
to commit, one thread succeeds with the insert but the second one fails with an
integrity error. When the second thread fails, the exception was uncaught and
led to the thread to die.

Now we look for this exception and in particular 1062 MySQL error code. If we do
hit it, we retry 3 times. Ideally, 2 should be sufficient since any subsequent
commit should be an update operation.

Integration tests will be updated to cover this case.

Testing done: In a DU without the UI workaround, tried to assign multiple roles
multiple times. Verified that the exceptions are logged in the log and the
subsequent retry succeeds.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/698)

Issues addressed: IAAS-615