Fix hostagent crash when RabbitMQ goes down.
This update detects when a RabbitMQ connection is closed. It will then
throw an error, and an outer loop will attempt to reconnect.

Testing done: Killed the rabbitmq-server process for both authorized and
unauthorized hosts. Ensured hostagent, bbmaster, and ostackhost did not
crash and were able to reconnect.

Issues Addressed: IAAS-644