Auto support bundle generation in backbone
- Add capability for capturing log files from the host. Currently, it is a very
simple file gathering code. It packs the logs into a tgz file
(/var/opt/pf9/hostagent/hostid/pf9-support.tgz). This file gets overwritten each
time the support bundle is generated.
- The hostagent shall automatically generate the support bundle and send it to
the master whenever the host goes in the 'failed' state from 'retrying' state
for the first time.
- The support file is then sent to the backbone master as a rabbitmq message.
The support file is sent down the wire as base64 encoded data. Since we use
pika, the data has to be sent out as a string.
- The backbone master receives this message, decodes it and writes it out to the
/opt/pf9/support/host-id directory. The file name is constructed as
hostname-timestamp.tgz.
- In case of error, the hostagent sends the error message to the master as part
of the message. The master logs this and continues.

Note that there is no REST API yet to initiate a support bundle generation for a
host.

Testing done: Patched DU and host with these changes and simulated failure cases
to verify that the log bundle was generated.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/714/)

Issues addressed: IAAS-637