Fix 70-user-setup.py customization script
The script was opening and appending project details to the end of the nova and
janitor ini files. This can cause an issue if this step was run more than once,
leading to multiple entries. Instead, use the openstack-config (crudini) script
to set these values up. It will now update the existing settings instead of
adding the block again. See IAAS-757 for more details.

Testing done: Ran customization and it works fine.

Reviewed by: rparikh, rdeuel (https://rbcommons.com/s/platform9/r/811/)

Issues addressed: IAAS-757