Move pf9 user home dir and login shell
Make the pf9 user login as /sbin/nologin (cannot login as pf9 user) and the / as
its home directory.

Note that the test machines today have a builtin pf9 user. So this code path is
not fully exercised. I am working on it separately since it needs changes in
multiple places in pf9lab and integration tests.

Testing done: Created a new DU with this change. On the host to be authorized,
removed the pf9 user and group. Installed hostagent and authorized both roles.
All succeeded.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/837/)

Issues addressed:IAAS-692