Automatically set active version of a role in resource manager
On upgrade, if a role version is active in the resmgr DB we continued to leave
that version as active in the DB. The intent was the du_ctl script would switch
the active version at a later point in time. However, this was based on the
assumption that multiple versions of the role would coexist in the DU (for
quicker rollback, etc.). The current implementation doesn't support such
multiple version setup and the current implementation is causing issues after
the role was upgraded in the DU.

With this change, resmgr will treat the metadata files as the source of truth
for active versions. If the metadata indicates that the role was new, we insert
it into the DB and mark its version as active. If the metadata indicates that
the role existed, then we mark all other versions of the role (if any) as
inactive and mark this role version as active. If the metadata was missing some
role which exists in the DB (which theoretically should not happen), these DB
entries are left untouched.

Considering this behavior change, I removed all the code which is unused now
including the REST API to set the active version.

The du_ctl script also should be updated, but that needs to wait since we need
this capability to update one more customer to Alpha2.

Testing done: Patched a DU and manually tried following scenarios:
1. New roles were added correctly into the DB when the DB was scrubbed.
2. Updated a role rpm in the DB to a new version and ensured on restart of
resource manager, the new version is marked as active and others are marked as
inactive.
3. Restarted resource manager with no changes to role versions and ensured that
active roles in the DB remain consistent from before.

Reviewed by: leb (https://rbcommons.com/s/platform9/r/841/)

Issues addressed: IAAS-780