Additional case to generate support bundle when convergence fails in hostagent
Previously, the support bundle was generated when the convergence attempt in
hostagent failed during install, configure or set service state of a pf9 app.
There are cases when the hostagent succeeds in all those cases but the app could
hit an exception and crash. We need to go to the customer in such cases and ask
for logs.

With this change, we will generate the support bundle when we fail in such cases
as well. The generation of support bundle has been moved now to the last
convergence
attempt (before the actual convergence is attempted). This will handle both
cases
when the convergence itself fails and when the app crashes because of an
exception.

Testing done: Modified hostagent.conf to have short retry period and attempts.
Setup ostackhost role but killed ostackhost with 'watch -n 5 pkill nova-compute'
to ensure ostackhost went down after hostagent started it each time. Ensured the
support bundle was generated only once and uploaded to the DU.

Edited the config of ostackhost to fail on each convergence setting itself and
ensured that there is no regression in the previous support bundle generation
logic.

Reviewed by: leb(https://rbcommons.com/s/platform9/r/846/)

Issues addressed: IAAS-791